cpu/hir: fix unsigned MulHi sign-extension on the ARM64 const-fold path
Port xenia-edge 643c13668 (has207). In Value::MulHi, the unsigned INT64
case on the non-MSVC (#else = GCC/Clang -> our Android/ARM64) path cast
the signed int64_t storage straight to unsigned __int128:
    static_cast<unsigned __int128>(constant.i64)
which SIGN-extends a high-bit-set value (top 64 bits become all ones)
instead of zero-extending. That corrupts constant-folded unsigned
multiply-high (mulhwu/mulhdu) whenever an operand has bit 63 set -
common in hashes, RNG, and fixed-point math. This is a live wrong-result
bug on the Thor (we build the #else path).

Fix: zero-extend through uint64_t before widening, matching the already-
correct INT32 case just above. Pure integer correctness, default-on,
no cvar. Build-verified (assembleGithubDebug, arm64 ndk-build OK).

+++ src/xenia/cpu/hir/value.cc
@@ -410,9 +410,15 @@ void Value::MulHi(Value* other, bool is_unsigned) {
       }
 #else
       if (is_unsigned) {
+        // Zero-extend (not sign-extend) the signed int64_t storage before
+        // widening: casting constant.i64 straight to unsigned __int128
+        // sign-extends a high-bit-set value, corrupting unsigned mulhdu/mulhwu
+        // constant folding on the GCC/Clang (Android/ARM64) path. Mirrors the
+        // already-correct INT32 case above. Ported from xenia-edge 643c13668.
         constant.i64 = static_cast<uint64_t>(
-            (static_cast<unsigned __int128>(constant.i64) *
-             static_cast<unsigned __int128>(other->constant.i64)) >>
+            (static_cast<unsigned __int128>(static_cast<uint64_t>(constant.i64)) *
+             static_cast<unsigned __int128>(
+                 static_cast<uint64_t>(other->constant.i64))) >>
             64);
       } else {
         constant.i64 = static_cast<uint64_t>(
